requirements.txt
blessings==1.7
certifi==2024.2.2
charset-normalizer==2.0.12
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
einops==0.4.1
filelock==3.4.1
gpustat==0.6.0
huggingface-hub==0.4.0
idna==3.7
importlib-metadata==4.8.3
importlib-resources==5.4.0
joblib==1.1.1
jsonpatch==1.32
jsonpointer==2.3
kiwisolver==1.3.1
matplotlib==3.3.4
networkx==2.5.1
numpy==1.19.5
nvidia-ml-py3==7.352.0
packaging==21.3
Pillow==8.4.0
psutil==5.9.8
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.27.1
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
timm==0.6.12
torch==1.7.1+cu110
torchaudio==0.7.2
torchvision==0.8.2+cu110
tornado==6.1
tqdm==4.64.1
typing_extensions==4.1.1
urllib3==1.26.18
visdom==0.2.4
websocket-client==1.3.1
zipp==3.6.0